style(banks-user): move the BanksUser interface in a specific file
