+++ src/lib/index.ts
@@ -4,3 +4,4 @@ export * from './Algoan.dto';
 export * from './Application.interface';
 export * from './Folder.interface';
 export * from './Logger.interface';
+export * from './BanksUser.interface';
